+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/properties/TraceabilityProperties.java
@@ -54,5 +54,6 @@ public class TraceabilityProperties {
 
     private String irsBase;
     private String submodelBase;
+    private String discoveryType;
 
 }