+++ charts/traceability-foss/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
+### Changed
+
+- #50 added the discovery type configurable, with a default value of bpnl under Traceability (DISCOVERY_TYPE).
+
 ## [1.3.44] - 2024-08-16
 ### No changes